change type to long int

+++ 008-largest_product_in_a_series/largest_product_in_a_series.hpp
@@ -1,3 +1,3 @@
 #include <string>
 
-int LargestProduct(std::string n, int m);
+long int LargestProduct(std::string n, int m);
